+++ src/main/java/online/kingdomkeys/kingdomkeys/handler/InputHandler.java
@@ -60,7 +60,6 @@ public class InputHandler {
     @Nullable public List<UUID> portalCommands;
     @Nullable public List<Member> targetsList;
     @Nullable public List<Limit> limitsList;
-    @Nullable public List<String> magicList;
     @Nullable public Map<Integer, ItemStack> itemsList;
     @Nullable public LinkedHashMap<String, Integer> reactionList;
     
@@ -475,7 +474,6 @@ public class InputHandler {
         if(playerData != null && worldData != null) {
             //this.magicsMap = Utils.getSortedMagics(playerData.getMagicsMap());
             this.portalCommands = worldData.getAllPortalsFromOwnerID(player.getUUID());
-            this.magicList = ModConfigs.magicDisplayedInCommandMenu.stream().filter(magic -> playerData.getMagicsCastMap().containsKey(magic)).toList();
             this.limitsList = Utils.getSortedLimits(Utils.getPlayerLimitAttacks(player));
 
             if(worldData.getPartyFromMember(player.getUUID()) != null) {